review template service to use the templates object returned by the api update and refactor linkto
