+++ components/LinkTo/index.js
@@ -1,0 +1,2 @@
+export * from "./LinkTo"
+export * from "./properties"
